testing/web-platform/tests/resource-timing/buffer-full-add-entries-during-callback-that-drop.html

<!DOCTYPE HTML>
<html>
<head onload>
<meta charset="utf-8" />
<title>This test validates that synchronously adding entries in onresourcetimingbufferfull callback results in these entries being properly handled.</title>
<link rel="author" title="Intel" href="http://www.intel.com/" />
<link rel="help" href="http://www.w3.org/TR/resource-timing/#performanceresourcetiming"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/append.js"></script>
</head>
<body>
<script>
promise_test(()=>{
    return new Promise((resolve, reject)=>{
        const resource_timing_buffer_size = 1;
        performance.clearResourceTimings();
        var add_entry = function() {
            performance.setResourceTimingBufferSize(resource_timing_buffer_size + 1);
            // The sync attribute is added to the secondary buffer, so will be last one there and eventually dropped.
            xhrScript("resources/empty.js?xhr");
            resolve();
        }
        performance.addEventListener('resourcetimingbufferfull', add_entry);
        performance.setResourceTimingBufferSize(resource_timing_buffer_size);
        // Scripts appended in JS to ensure setResourceTimingBufferSize is called before.
        // This resource gets buffered in the resource timing entry buffer.
        appendScript('resources/empty.js');
        // This resource overflows the entry buffer, and goes into the secondary buffer.
        appendScript('resources/empty_script.js');
    });
}, "Prepare test");
promise_test(function() {
    return new Promise(function(resolve, reject) {
        let waitForIt = function() {
            resolve();
        }
        step_timeout(waitForIt, 0);
    });
}, "Wait for next task");
promise_test(()=>{
    return new Promise((resolve, reject)=>{
        let entries = performance.getEntriesByType('resource');
        assert_equals(entries.length, 2,
                      'Both entries should be stored in resource timing buffer since its increases size once it overflows.');
        assert_true(entries[0].name.includes('empty.js'), "empty.js is in the entries buffer");
        assert_true(entries[1].name.includes('empty_script.js'), "empty_script.js is in the entries buffer");
        resolve();
    });
}, "Test");
</script>
</body>
</html>
